se cambia el tipo de dato al precio y al total

+++ common/src/main/java/com/social/common/event/PurchaseEvent.java
@@ -14,7 +14,7 @@ public class PurchaseEvent {
     private String productId;
     private String userId;
     private Integer quantity;
-    private Double total;
+    private float total;
 
     @Override
     public String toString() {
